move project template to extension
