+++ src/SimpleAccounting/Abstractions/IFileSystem.cs
@@ -6,7 +6,6 @@ namespace lg2de.SimpleAccounting.Abstractions
 {
     using System;
     using System.Collections.Generic;
-    using System.IO;
 
     /// <summary>
     ///     Defines abstraction for file system access not available in .NET framework.
@@ -61,7 +60,7 @@ namespace lg2de.SimpleAccounting.Abstractions
         /// <summary>
         ///     Retrieves the drive names of all logical drives on a computer.
         /// </summary>
-        /// <returns>An enumeration of type <see cref="DriveInfo"/> that represents the logical drives on a computer.</returns>
+        /// <returns>An enumeration of type <see cref="System.IO.DriveInfo"/> that represents the logical drives on a computer.</returns>
         IEnumerable<(string RootPath, string Format)> GetDrives();
     }
 }
